requirements.txt
tensorflow
opencv-python
mediapipe
scikit-learn
numpy
Pillow
mlflow
fastapi
uvicorn[standard]
python-multipart
requests